% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{somar}
\alias{somar}
\title{function to sum(titulo)}
\usage{
somar(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{It is another number.}
}
\description{
This function is good to sum.
}
\examples{
somar(2, 3)
5

}
