% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{subtract}
\alias{subtract}
\title{function to subtract}
\usage{
subtract(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{It is another number.}
}
\description{
This function is good to subtract.
}
\examples{
subtract(2, 3)
-1

}
